fix(monobank): widen invoice amount fields from int to long
MonobankInvoiceRequest.Amount was int, capping invoices at ~21.5M UAH in kopecks. Widens it plus the response-side MonobankInvoiceStatus.Amount/FinalAmount and MonobankCancelListItem.Amount to long, matching the webhook DTO which already used long. Drops the now-lossy (int) casts at the two invoice-creation sites and in the stub client. Order/Refund DB columns stay int (per-order prices cannot approach the limit).

+++ backend/src/FuelFlow.API/Features/Orders/SharedServices/Monobank/Models/MonobankInvoiceRequest.cs
@@ -2,7 +2,8 @@ namespace FuelFlow.API.Features.Orders.SharedServices.Monobank.Models;
 
 public sealed class MonobankInvoiceRequest
 {
-    public int Amount { get; set; }
+    // Kopecks. long so bulk invoices cannot overflow int (~21.5M UAH).
+    public long Amount { get; set; }
     public string Currency { get; set; } = "UAH";
     public string? WebhookUrl { get; set; }
     public string? RedirectUrl { get; set; }
